+++ apps/web-tools/CHANGELOG.md
@@ -16,6 +16,13 @@
   - @ledgerhq/hw-ledger-key-ring-protocol@0.5.4-next.0
   - @ledgerhq/ledger-key-ring-protocol@0.7.1-next.0
 
+## 0.25.1-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies [[`51bfea9`](https://github.com/LedgerHQ/ledger-live/commit/51bfea93d2b52ab552bb7f4932bfd225134f3238)]:
+  - @ledgerhq/live-common@34.52.1-hotfix.0
+
 ## 0.25.0
 
 ### Minor Changes