+++ e2e/desktop/CHANGELOG.md
@@ -18,6 +18,22 @@
   - @ledgerhq/live-env@2.21.0-next.0
   - @ledgerhq/ledger-key-ring-protocol@0.7.1-next.0
 
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies [[`51bfea9`](https://github.com/LedgerHQ/ledger-live/commit/51bfea93d2b52ab552bb7f4932bfd225134f3238)]:
+  - @ledgerhq/live-common@34.52.1
+  - @ledgerhq/live-cli@24.27.1
+
+## 0.9.1-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies [[`51bfea9`](https://github.com/LedgerHQ/ledger-live/commit/51bfea93d2b52ab552bb7f4932bfd225134f3238)]:
+  - @ledgerhq/live-common@34.52.1-hotfix.0
+  - @ledgerhq/live-cli@24.27.1-hotfix.0
+
 ## 0.9.0
 
 ### Minor Changes