+++ libs/coin-tester-modules/coin-tester-evm/CHANGELOG.md
@@ -17,6 +17,20 @@
   - @ledgerhq/coin-tester@0.12.0-next.0
   - @ledgerhq/live-signer-evm@0.10.1-next.0
 
+## 1.10.1
+
+### Patch Changes
+
+- Updated dependencies [[`51bfea9`](https://github.com/LedgerHQ/ledger-live/commit/51bfea93d2b52ab552bb7f4932bfd225134f3238)]:
+  - @ledgerhq/live-common@34.52.1
+
+## 1.10.1-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies [[`51bfea9`](https://github.com/LedgerHQ/ledger-live/commit/51bfea93d2b52ab552bb7f4932bfd225134f3238)]:
+  - @ledgerhq/live-common@34.52.1-hotfix.0
+
 ## 1.10.0
 
 ### Minor Changes